[[with_stacktrace]] reimplementation
Rather than allocate heap space (next to the exception object?), we
use a thread_local with_stacktrace object and move-construct the catch
object immediately on entry into the catch block.

Builds on recent std::stacktrace support, added in https://gcc.gnu.org/pipermail/gcc-patches/2022-January/588617.html
Pass --enable-libstdcxx-backtrace=yes to configure and link libstdc++_libbacktrace.a

We pass the caught type through the exception machinery in a special
WITH_STACKTRACE_TYPE type wrapper.  This is mangled as if it was an
attribute (we can't use an actual attribute-modified type, because (a)
attributes can't variably modify tagged types, and (b) only platform
attributes are allowed to affect type identity).

The function to perform the work of collecting the stacktrace at throw
point is not compiled into libsupc++ but instead grabbed out of
namespace std when doing rtti for the catch block.  This is more
flexible and in keeping with zero-overhead principle.