Only allow functional-style cast to perform direct
-initialization

For now, this is a hard error, only in C++23 mode.

Rationale:

https://cplusplus.github.io/LWG/issue3528 is a welcome fix for an
insidious bug (that we have been bit by).

However, the resolution (requiring is_constructible i.e. for
direct-initialization to be well-formed) goes beyond the justification
for the issue (since there are conversions allowed by static_cast that
are not available to direct-initialization), leaves open the possibility
that there could be code where direct-initialization is well-formed but
has *different* semantics to those selected by cast notation, and leaves
users vulnerable to committing the same error in their libraries.

For example, `static_cast<int&&>(declval<int&>())` is valid, but `int&&
t(declval<int&>());` is not; likewise base-to-derived pointer and
reference conversions, enumeration conversions, and so forth.

Instead, we propose deprecating the cast notation interpretation of the
functional notation entirely; that is, we wish to strike the first
sentence of [expr.type.conv]/2.  Per this change, functional notation
will be fully consistent with direct-initialization, except that it will
also be able to construct void prvalues (but only with an initializer of
() or {}).

This patch implements said deprecation as a hard error in C++23 mode,
and fixes compiler, library and testcase code as necessary by replacing
offending functional notation with the weakest possible cast, excepting
that in some testcases it has not been possible to determine this so the
cast notation is used.

https://eel.is/c++draft/expr.compound#expr.type.conv-2.sentence-1